+++ Changelog.md
@@ -7,6 +7,10 @@ Enhancements
     * intended for extension libs
     * restores rspec-1 behavior
 
+Bug fixes
+
+* Properly parse `SPEC_OPTS` options.
+
 ### 2.9.0 / 2012-03-17
 [full changelog](http://github.com/rspec/rspec-core/compare/v2.8.0...v2.9.0)